user_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.4.2
mysql-connector-python==8.2.0